+++ lib/components/base_component.dart
@@ -50,6 +50,7 @@ abstract class BaseComponent extends Component {
   @override
   void update(double dt) {
     _effectsHandler.update(dt);
+    _children.forEach((c) => c.update(dt));
   }
 
   @mustCallSuper
@@ -62,20 +63,11 @@ abstract class BaseComponent extends Component {
     }
 
     canvas.save();
-    _children.forEach((c) => _renderChild(canvas, c));
+    _children.forEach((c) => c.render(canvas));
     canvas.restore();
   }
 
   void renderDebugMode(Canvas canvas) {}
-
-  void _renderChild(Canvas canvas, Component c) {
-    if (!c.loaded) {
-      return;
-    }
-    c.render(canvas);
-    canvas.restore();
-    canvas.save();
-  }
 
   @protected
   void prepareCanvas(Canvas canvas) {}
@@ -111,12 +103,12 @@ abstract class BaseComponent extends Component {
   List<ComponentEffect> get effects => _effectsHandler.effects;
 
   /// Uses the game passed in, or uses the game from [HasGameRef] otherwise,
-  /// to prepare the child component before it is added to the list of children
+  /// to prepare the child component before it is added to the list of children.
+  /// Note that this component needs to be added to the game first if
+  /// [this.gameRef] should be used to prepare the child.
+  /// For children that don't need preparation from the game instance can
+  /// disregard both the options given above.
   void addChild(Component c, {Game gameRef}) {
-    assert(
-      gameRef != null || this is HasGameRef,
-      "Need gameRef either as an argument or from the HasGameRef mixin",
-    );
     gameRef ??= (this as HasGameRef).gameRef;
     if (gameRef is BaseGame) {
       gameRef.prepare(c);