+++ test/components/composed_component_test.dart
@@ -9,10 +9,26 @@ import 'package:flame/extensions/vector2.dart';
 import 'package:flutter/gestures.dart';
 import 'package:test/test.dart';
 
+import '../util/mock_canvas.dart';
+
 class MyGame extends BaseGame with HasTapableComponents {}
 
 class MyTap extends PositionComponent with Tapable, Resizable {
   bool tapped = false;
+  bool updated = false;
+  bool rendered = false;
+
+  @override
+  void update(double dt) {
+    super.update(dt);
+    updated = true;
+  }
+
+  @override
+  void render(Canvas canvas) {
+    super.render(canvas);
+    rendered = true;
+  }
 
   @override
   bool onTapDown(TapDownDetails details) {
@@ -38,15 +54,30 @@ void main() {
     test('taps and resizes children', () {
       final MyGame game = MyGame();
       final MyTap child = MyTap();
-      final MyComposed wrapper = MyComposed()..addChild(child);
+      final MyComposed wrapper = MyComposed();
 
       game.size = size;
+      wrapper.addChild(child);
       game.add(wrapper);
       game.update(0.0);
       game.onTapDown(1, TapDownDetails(globalPosition: const Offset(0.0, 0.0)));
 
       expect(child.gameSize, size);
       expect(child.tapped, true);
+    });
+
+    test('updates and renders children', () {
+      final MyGame game = MyGame();
+      final MyTap child = MyTap();
+      final MyComposed wrapper = MyComposed();
+
+      wrapper.addChild(child);
+      game.add(wrapper);
+      game.update(0.0);
+      game.render(MockCanvas());
+
+      expect(child.rendered, true);
+      expect(child.updated, true);
     });
   });
 }